requirements.txt
aiofiles
aiohttp
apscheduler
beautifulsoup4
deepai
edge-tts
ffmpeg-python
flask
future
gitpython
google-api-python-client
gpytranslate
gtts
heroku3
httpx[http2]==0.28.1
pyrofork
lexica-api
motor
nekosbest
numpy
opencv-python
pillow==12.1.1
psutil
ntgcalls==2.1.0
py-tgcalls==2.2.11
pycryptodome
pycryptodomex
pydub
pyfiglet
pyshorteners
python-dotenv
python-whois
git+https://github.com/CertifiedDevloper/youtube-search-python
pyyaml
qrcode
requests
speedtest-cli
SpeechRecognition
spotipy
telegraph
tgcrypto
unidecode
wget
yt-dlp[default] @ git+https://github.com/yt-dlp/yt-dlp.git
nudenet
